backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
python-docx==1.1.0
python-pptx==0.6.23
google-generativeai==0.3.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
